% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{df2file}
\alias{df2file}
\title{Convert Data Frame to Valid JSON Format for Serving}
\usage{
df2file(df, output = "./data/df.json", ...)
}
\arguments{
\item{df}{A data frame.}

\item{output}{A string. The path to the exported data frame.}

\item{...}{Additional arguments arguments passed on to
\link[jsonlite]{toJSON}.}
}
\description{
Convert Data Frame to Valid JSON Format for Serving
}
