% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{from_repo}
\alias{from_repo}
\title{Build URL to a file from GitHub Repository}
\usage{
from_repo(username, repo, branch = "master", path)
}
\arguments{
\item{username}{A string. The owner of the repo.}

\item{repo}{A string. The name of the repo.}

\item{branch}{A string. The branch the file is on.}

\item{path}{A string. The path to the file.}
}
\value{
An URL to the path.
}
\description{
This function builds a URL in the form:
\code{https://raw.githubusercontent.com/<username>/<repo>/<branch>/<path>}.
}
